refactor(pools): reorganize base contract and base provider
