+++ src/base-classes/index.ts
@@ -1,0 +1,2 @@
+export * from './base-contract'
+export * from './base-provider'
