+++ src/pools/assisted-provider.ts
@@ -1,5 +1,5 @@
 import { miscErc20CollateralPool } from '../artifacts'
-import { BaseProvider } from '../base-classes/base-provider'
+import { BaseProvider } from '../base-classes'
 import {
   Abi,
   ERC20CollateralPoolConstructorParams,
