+++ src/utilities/erc20.ts
@@ -1,7 +1,7 @@
 import { ethers } from 'ethers'
 
 import { miscErc20 } from '../artifacts'
-import { BaseContract } from '../base-classes/base-contract'
+import { BaseContract } from '../base-classes'
 import { ecpErrorMessage } from '../errors'
 import { Abi, PrivateKey } from '../types/types'
